+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+### 1.0.2
+- [x] Fixed Black Temple tracking
+
 ### 1.0.1
 - [x] Fixed some Worldbosses typos
 
 ### 1.0.0
 - [x] Initial fork commit
-- [x] Fixed Sunwell, SSC and Tempest Keep instances
+- [x] Fixed Sunwell, SSC and Tempest Keep tracking
 - [x] Added LFR tracking
 - [x] Added Worldbosses tracking